use float64 for latencies, uint64 is not large enough

+++ collector/sys_user_summary.go
@@ -16,6 +16,7 @@ package collector
 import (
 	"context"
 	"database/sql"
+
 	"github.com/go-kit/log"
 	"github.com/prometheus/client_golang/prometheus"
 )
@@ -109,10 +110,10 @@ func (ScrapeSysUserSummary) Scrape(ctx context.Context, db *sql.DB, ch chan<- pr
 	var (
 		user                   string
 		statements             uint64
-		statement_latency      uint64
+		statement_latency      float64
 		table_scans            uint64
 		file_ios               uint64
-		file_io_latency        uint64
+		file_io_latency        float64
 		current_connections    uint64
 		total_connections      uint64
 		unique_hosts           uint64
